requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.2
opencv-python>=4.7.0
tqdm>=4.65.0
tensorflow>=2.12.0
Pillow>=10.0.0
scikit-learn>=1.2.0